server/requirements.txt
fastapi-slim
python-socketio
uvicorn
boto3
python-dotenv